(refactor) Move hooks into their own directory
