+++ src/components/form-editor/element-editor/element-editor.tsx
@@ -1,6 +1,6 @@
 import React, { useContext, useEffect, useState } from "react";
 import { Accordion, Button, Column, Row, TextInput } from "@carbon/react";
-import { Schema } from "../../../api/types";
+import { Schema } from "../../../types";
 import { SchemaContext } from "../../../context/context";
 import PageElement from "./elements/page";
 import styles from "./element-editor.scss";
