+++ src/components/form-editor/element-editor/elements/question.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { TrashCan } from "@carbon/react/icons";
 import { Button, Column, Row } from "@carbon/react";
-import { Question } from "../../../../api/types";
+import { Question } from "../../../../types";
 import EditQuestion from "../../modals/edit-question";
 import styles from "./elements.scss";
