+++ src/components/form-editor/modals/new-section.tsx
@@ -12,7 +12,7 @@ import {
   SelectItem,
   TextInput,
 } from "@carbon/react";
-import { Section } from "../../../api/types";
+import { Section } from "../../../types";
 import { Add } from "@carbon/react/icons";
 import { SchemaContext } from "../../../context/context";
 import { showToast } from "@openmrs/esm-framework";
