+++ src/hooks/useConceptLookup.ts
@@ -1,8 +1,8 @@
 import useSWR from "swr";
 import { openmrsFetch, showToast } from "@openmrs/esm-framework";
-import { Concept } from "./types";
+import { Concept } from "../types";
 
-export function useSearchConcept(conceptID) {
+export function useConceptLookup(conceptID) {
   const url =
     conceptID != "" ? `/ws/rest/v1/concept?q=${conceptID}&v=full` : null;
   const { data, error } = useSWR<{ data: { results: Array<Concept> } }, Error>(
