+++ src/hooks/usePocForms.ts
@@ -1,8 +1,8 @@
 import useSWR from "swr";
 import { openmrsFetch } from "@openmrs/esm-framework";
-import { Form } from "./types";
+import { Form } from "../types";
 
-export function usePOCForms() {
+export function usePocForms() {
   const url =
     "/ws/rest/v1/form?q=POC&v=custom:(uuid,name,encounterType:(uuid,name),version,published,retired,resources:(uuid,name,dataType,valueReference))";
